hw3/inference/inference-requirements.txt
numpy>=1.19.5
flask>=2.0.1
pillow>=8.3.1
werkzeug>=2.0.1
gunicorn>=20.1.0